Move Geolocation tests into their own file
Follow-up to the previous commit. test/lib/validator/common/
common_utils_test.rb held three geolocation tests
(test_format_insdc_latlon, test_convert_latlon_insdc2iso,
test_country_at) that had nothing to do with CommonUtils anymore now
that Geolocation is its own module.

Move them to test/lib/validator/common/geolocation_test.rb under a new
TestGeolocation class. The new file has no setup block — the tests
don't need @@null_accepted / @@eutils_api_key, so the setup that
common_utils_test.rb needs is now only built for CommonUtils tests.

No assertion changes. Counts unchanged: 324 / 2600 / 0 failures.
